+++ tests/Integration/GeolocationTest.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace CrowdSecBouncer\Tests\Integration;
 
-use CrowdSecBouncer\StandaloneBouncer;
+use CrowdSec\LapiClient\Bouncer as BouncerClient;
+use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
+use CrowdSec\RemediationEngine\LapiRemediation;
+use CrowdSecBouncer\AbstractBouncer;
 use CrowdSecBouncer\Constants;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
+
 /**
- * @covers \CrowdSecBouncer\StandaloneBouncer::clearCache
  * @covers \CrowdSecBouncer\AbstractBouncer::getRemediationForIp
  *
  * @uses \CrowdSecBouncer\AbstractBouncer::__construct
@@ -29,8 +32,7 @@ use Psr\Log\LoggerInterface;
  * @uses \CrowdSecBouncer\Configuration::addTemplateNodes
  * @uses \CrowdSecBouncer\Configuration::cleanConfigs
  * @uses \CrowdSecBouncer\Configuration::getConfigTreeBuilder
- * @uses \CrowdSecBouncer\StandaloneBouncer::__construct
- * @uses \CrowdSecBouncer\StandaloneBouncer::handleTrustedIpsConfig
+ * @uses \CrowdSecBouncer\AbstractBouncer::clearCache
  */
 final class GeolocationTest extends TestCase
 {
@@ -39,7 +41,7 @@ final class GeolocationTest extends TestCase
 
     /** @var LoggerInterface */
     private $logger;
-    /** @var bool  */
+    /** @var bool */
     private $useCurl;
     /** @var bool */
     private $useTls;
@@ -122,13 +124,16 @@ final class GeolocationTest extends TestCase
             'use_curl' => $this->useCurl,
             'cache_system' => Constants::CACHE_SYSTEM_PHPFS,
             'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR,
-            'stream_mode' => false
+            'stream_mode' => false,
         ];
 
-        $bouncer = new StandaloneBouncer($bouncerConfigs, $this->logger);
+        $client = new BouncerClient($bouncerConfigs);
+        $cache = new PhpFiles($bouncerConfigs);
+        $lapiRemediation = new LapiRemediation($bouncerConfigs, $client, $cache);
+
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncer::class, [$bouncerConfigs, $lapiRemediation]);
 
         $bouncer->clearCache();
-
 
         $this->assertEquals(
             'captcha',
@@ -145,7 +150,11 @@ final class GeolocationTest extends TestCase
         // Disable Geolocation feature
         $geolocationConfig['enabled'] = false;
         $bouncerConfigs['geolocation'] = $geolocationConfig;
-        $bouncer = new StandaloneBouncer($bouncerConfigs, $this->logger);
+        $client = new BouncerClient($bouncerConfigs);
+        $cache = new PhpFiles($bouncerConfigs);
+        $lapiRemediation = new LapiRemediation($bouncerConfigs, $client, $cache);
+
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncer::class, [$bouncerConfigs, $lapiRemediation]);
         $bouncer->clearCache();
 
         $this->assertEquals(
@@ -158,7 +167,10 @@ final class GeolocationTest extends TestCase
         $this->watcherClient->setSecondState();
         $geolocationConfig['enabled'] = true;
         $bouncerConfigs['geolocation'] = $geolocationConfig;
-        $bouncer = new StandaloneBouncer($bouncerConfigs, $this->logger);
+        $client = new BouncerClient($bouncerConfigs);
+        $cache = new PhpFiles($bouncerConfigs);
+        $lapiRemediation = new LapiRemediation($bouncerConfigs, $client, $cache);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncer::class, [$bouncerConfigs, $lapiRemediation]);
         $bouncer->clearCache();
 
         $this->assertEquals(
@@ -176,6 +188,7 @@ final class GeolocationTest extends TestCase
 
     /**
      * @group integration
+     *
      * @dataProvider maxmindConfigProvider
      *
      * @throws \Symfony\Component\Cache\Exception\CacheException|\Psr\Cache\InvalidArgumentException
@@ -193,11 +206,15 @@ final class GeolocationTest extends TestCase
             'geolocation' => $geolocationConfig,
             'use_curl' => $this->useCurl,
             'cache_system' => Constants::CACHE_SYSTEM_PHPFS,
-            'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR
+            'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR,
         ];
 
-        $bouncer = new StandaloneBouncer($bouncerConfigs, $this->logger);
-        $cacheAdapter= $bouncer->getRemediationEngine()->getCacheStorage();
+        $client = new BouncerClient($bouncerConfigs);
+        $cache = new PhpFiles($bouncerConfigs);
+        $lapiRemediation = new LapiRemediation($bouncerConfigs, $client, $cache);
+
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncer::class, [$bouncerConfigs, $lapiRemediation]);
+        $cacheAdapter = $bouncer->getRemediationEngine()->getCacheStorage();
         $cacheAdapter->clear();
 
         // Warm BlockList cache up