+++ tests/end-to-end/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    preset: "jest-playwright-preset",
-    testRunner: "jest-circus/runner",
-    testEnvironment: "./CustomEnvironment.js",
-    testSequencer: "./testSequencer.js",
-    setupFilesAfterEnv: ["expect-playwright"],
-};